Complete httpd server for pj3 need
A basic http server is constructed, supporting http GET method only.

This server is not secure in many ways though.
This server may leak out-of-scoped files to whoever connects it with
specially crafted pathname (use ./ to go to parent directory). And the
server tries to execute file if it has X pemission to the user running
that server.